2012 alignment recipe for CMSSW_5_3_16_patch1
This commit includes the latest alignment developments, for example the code
for Lorentz-angle and backplane corrections, which has not been added to the
official releases. It is a digest of the setup in
MPproduction/CMSSW_5_3_7_patch1_nbartosi

The code is equivalent to the following CVS tags:
  V03-12-05            Alignment/CommonAlignment
  V04-00-09-53X-calib-V01  Alignment/CommonAlignmentAlgorithm
  V01-10-05-02         Alignment/CommonAlignmentMonitor
  V01-13-00-01         Alignment/CommonAlignmentProducer
  V00-29-04-02         Alignment/MillePedeAlignmentAlgorithm
  V01-05-04            Alignment/ReferenceTrajectories
  V02-07-02            Alignment/TrackerAlignment
  lq_53backport_BP_v1  CalibTracker/Configuration
  lq_53backport_BP_v1  CalibTracker/Records
  lq_53backport_BP_v1  CalibTracker/SiStripESProducers
  lq_53backport_BP_v1  CondCore/SiStripPlugins
  lq_53backport_BP_v1  CondFormats/DataRecord
  V18-00-15            CondFormats/SiStripObjects
  lq_53backport_BP_v2  RecoLocalTracker/Records
  ls1_current          RecoLocalTracker/SiPixelRecHits
  V06-00-03_withSiStripBackplaneCorr  RecoLocalTracker/SiStripRecHitConverter

On top, the following manual changes are applied:

1) fix of compilation errors due to missing libraries:

  rm RecoLocalTracker/SiPixelRecHits/interface/VVIObjF.h
  rm RecoLocalTracker/SiPixelRecHits/src/VVIObjF.cc

2) additional, uncommitted changes in the following files taken
   from MPproduction/CMSSW_5_3_7_patch1_nbartosi):

  Alignment/CommonAlignmentProducer/interface/AlignmentTrackSelector.h
  Alignment/CommonAlignmentProducer/src/AlignmentTrackSelector.cc
  Alignment/CommonAlignmentProducer/python/AlignmentTrackSelector_cfi.py
  Alignment/CommonAlignmentProducer/bin/aligncond_split_iov.cc

3) configuring

  RecoLocalTracker/SiPixelRecHits/python/PixelCPETemplateReco_cfi.py

   such that mille-pede the implicit LA settings of the templates used in
   the pixel-hit reconstruction are taken into account by mille-pede. This
   assumes that the mille-pede is run consistently using the `template'
   method for pixel-hit reconstruction, i.e. that the setting
   `TTRHBuilder = "WithAngleAndTemplate"' is used for the track fitting.

To run the alignment, additionally the `Alignment/CommonAlignmentParametrization'
sub-package from the release has to be added using e.g.

  git cms-addpkg Alignment/CommonAlignmentParametrization

Even though this is exactly the version in the release, i.e. there are NO changes
applied to this sub-package, it has to be present (`sparse checked-out') in the
working release for reasons that are not fully clear to me. Otherwise, the mille
jobs will crash.

The described setup compiles, runs, and is being validated in `CMSSW_5_3_16_patch1'
in SL5 with SCRAM_ARCH=`slc5_amd64_gcc462'. (Releases before `patch1' do NOT
include the recommended hierarchy-constraints patch.)